+++ mycore-acl/src/main/java/org/mycore/mcr/acl/accesskey/restapi/v2/annotation/MCRRequireAccessKeyAuthorization.java
@@ -18,7 +18,7 @@
  * 59 Temple Place - Suite 330, Boston, MA  02111-1307 USA
  */
 
-package org.mycore.mcr.acl.accesskey.restapi.annotation;
+package org.mycore.mcr.acl.accesskey.restapi.v2.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
