+++ mycore-acl/src/main/java/org/mycore/mcr/acl/accesskey/restapi/v2/model/MCRAccessKeyInformation.java
@@ -18,7 +18,7 @@
  * 59 Temple Place - Suite 330, Boston, MA  02111-1307 USA
  */
 
-package org.mycore.mcr.acl.accesskey.restapi.model;
+package org.mycore.mcr.acl.accesskey.restapi.v2.model;
 
 import java.util.List;
