+++ mycore-acl/src/main/resources/components/acl/config/mycore.properties
@@ -5,12 +5,6 @@ MCR.CLI.Classes.External=%MCR.CLI.Classes.External%,org.mycore.mcr.acl.accesskey
 
 
 ##############################################################################
-# Strategy                                                                   #
-##############################################################################
-MCR.AccessKey.Strategy.ObjectTypes=mods
-
-
-##############################################################################
 # Configure Hibernate Mappings                                               #
 ##############################################################################
 MCR.Hibernate.Mappings=%MCR.Hibernate.Mappings%,org.mycore.mcr.acl.accesskey.backend.MCRAccessKey
@@ -26,7 +20,7 @@ MCR.EventHandler.MCRDerivate.011.Class=org.mycore.mcr.acl.accesskey.MCRAccessKey
 ##############################################################################
 # Jersey Resources                                                           #
 ##############################################################################
-MCR.RestAPI.V2.Resource.Packages=%MCR.Jersey.Resource.Packages%,org.mycore.mcr.acl.accesskey.restapi
+MCR.RestAPI.V2.Resource.Packages=%MCR.Jersey.Resource.Packages%,org.mycore.mcr.acl.accesskey.restapi.v2
 
 
 ##############################################################################
